+++ apm-sniffer/apm-sdk-plugin/dubbo-2.7.x-conflict-patch/src/main/java/org/apache/skywalking/apm/plugin/dubbo27/patch/MakeWrapperInterceptor.java
@@ -15,7 +15,7 @@
  *  limitations under the License.
  */
 
-package org.apache.skywalking.apm.plugin.dubbo.patch;
+package org.apache.skywalking.apm.plugin.dubbo27.patch;
 
 import org.apache.dubbo.common.bytecode.ClassGenerator;
 import org.apache.dubbo.common.bytecode.NoSuchPropertyException;
