+++ apm-sniffer/apm-sdk-plugin/dubbo-2.7.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/dubbo27/DubboInterceptor.java
@@ -17,7 +17,7 @@
  */
 
 
-package org.apache.skywalking.apm.plugin.dubbo;
+package org.apache.skywalking.apm.plugin.dubbo27;
 
 import org.apache.dubbo.common.URL;
 import org.apache.dubbo.rpc.Invocation;
